Move init part into common/__init__.py
