+++ nonebot_plugin_bing_chat/__init__.py
@@ -1,4 +1,4 @@
 from .common.utils import logger
 from .onebotv11.main import logger
 
-logger.warning('你使用的是开发版，可能会出现未知错误，请及时反馈')
+logger.opt(colors=True).warning('<red>你使用的是开发版，可能会出现未知错误，请及时反馈</red>')
